clean up for parser
