+++ dataplane-bin/Cargo.toml
@@ -1,0 +1,8 @@
+[package]
+name = "dataplane-bin"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+
+dataplane = { path = "../dataplane"}
