+++ dataplane-bin/src/main.rs
@@ -1,0 +1,5 @@
+extern crate dataplane;
+
+fn main() {
+    dataplane::real_main()
+}
