+++ dpdk-sys/Cargo.toml
@@ -6,6 +6,6 @@ description = "Low-level bindings to the Data Plane Development Kit (DPDK)"
 publish = false
 
 [build-dependencies]
-bindgen = { workspace = true, features = ["runtime", "experimental"] }
-doxygen-rs = { workspace = true }
+bindgen = { version = "0.71.1", features = ["runtime", "experimental"]}
+doxygen-rs = { version = "0.4.2" }
 dpdk-sysroot-helper = { path = "../dpdk-sysroot-helper", version = "0.0.1" }
