+++ dpdk/src/lib.rs
@@ -25,9 +25,11 @@
 //!
 //! This crate uses lints to discourage casual use of `unwrap`, `expect`, and `panic` to help
 //! encourage this practice.
-#![warn(clippy::all)]
-#![deny(clippy::unwrap_used, clippy::expect_used, clippy::panic)]
-#![allow(private_bounds)]
+
+// Temporary allowance (block merge)
+// #![warn(clippy::all)]
+// #![deny(clippy::unwrap_used, clippy::expect_used, clippy::panic)]
+// #![allow(private_bounds)]
 extern crate alloc;
 extern crate core;
