+++ net/src/parse.rs
@@ -1,5 +1,8 @@
+//! Packet parsing traits
+#![allow(missing_docs)] // temorary allowance (block merge)
+ 
 use std::num::NonZero;
-use std::ptr::NonNull;
+
 
 pub trait Parse: Sized {
     type Error: core::error::Error;
@@ -7,7 +10,7 @@ pub trait Parse: Sized {
 }
 
 pub trait DeParse {
-    type Error: core::error::Error;
+    type Error;
 
     fn size(&self) -> NonZero<usize>;
     fn write(&self, buf: &mut [u8]) -> Result<NonZero<usize>, DeParseError<Self::Error>>;
@@ -24,6 +27,17 @@ pub trait ParseWith {
         Self: Sized;
 }
 
+pub(crate) trait Step {
+    type Next;
+    fn step(&self, cursor: &mut Reader) -> Option<Self::Next>;
+}
+
+pub(crate) trait StepWith {
+    type Param;
+    type Next;
+    fn step_with(&self, param: &Self::Param, cursor: &mut Reader) -> Option<Self::Next>;
+}
+
 #[derive(thiserror::Error, Debug)]
 #[error("expected at least {expected} bytes, got {actual}")]
 pub struct LengthError {
@@ -32,62 +46,93 @@ pub struct LengthError {
 }
 
 #[derive(Debug)]
-#[repr(transparent)]
-pub(crate) struct Cursor {
-    pub(crate) inner: NonNull<[u8]>,
+pub(crate) struct Reader<'buf> {
+    pub(crate) inner: &'buf [u8],
+    pub(crate) remaining: usize,
 }
 
-impl Cursor {
-    #[allow(unsafe_code)] // TODO: explain how to use this type safely
-    pub(crate) fn new(buf: &[u8]) -> Cursor {
-        let range = buf.as_ptr_range();
-        let inner = NonNull::from(unsafe { core::slice::from_raw_parts(range.start, buf.len()) });
-        Cursor { inner }
+impl Reader<'_> {
+    fn remaining(&self) -> usize {
+        self.remaining
+    }
+}
+
+#[derive(Debug)]
+pub(crate) struct Writer<'buf> {
+    pub(crate) inner: &'buf mut [u8],
+    pub(crate) remaining: usize,
+}
+
+impl Reader<'_> {
+    
+    pub(crate) fn new(buf: &[u8]) -> Reader {
+        Reader {
+            inner: buf,
+            remaining: buf.len(),
+        }
     }
 
-    #[allow(unsafe_code)] // TODO: explain how to use this type safely
     fn consume(&mut self, n: NonZero<usize>) -> Result<(), LengthError> {
-        if n.get() >= self.inner.len() {
+        if n.get() >= self.remaining {
             return Err(LengthError {
                 expected: n,
-                actual: self.inner.len(),
+                actual: self.remaining,
             });
         };
-        let start = unsafe { self.inner.as_mut() }.as_mut_ptr();
-        let split = unsafe { start.byte_add(n.get()) };
-        let inner = NonNull::from(unsafe {
-            core::slice::from_raw_parts_mut(split, self.inner.len() - n.get())
-        });
-        self.inner = inner;
+        self.remaining -= n.get();
         Ok(())
     }
 
-    #[allow(unsafe_code)] // TODO: document safe usage
     pub(crate) fn parse<T: Parse>(&mut self) -> Result<(T, NonZero<usize>), ParseError<T::Error>> {
-        let (value, len_consumed) = T::parse(unsafe { self.inner.as_mut() })?;
+        let current = self.inner.len() - self.remaining;
+        let (value, len_consumed) = T::parse(&self.inner[current..])?;
         match self.consume(len_consumed) {
             Ok(()) => Ok((value, len_consumed)),
             Err(e) => Err(ParseError::LengthError(e)),
         }
     }
 
-    #[allow(unsafe_code)] // TODO: document safe usage
     pub(crate) fn parse_with<T: ParseWith>(
         &mut self,
         param: <T as ParseWith>::Param,
     ) -> Result<(T, NonZero<usize>), ParseError<T::Error>> {
-        let (value, len_consumed) = T::parse_with(param, unsafe { self.inner.as_mut() })?;
+        let current = self.inner.len() - self.remaining;
+        let (value, len_consumed) = T::parse_with(param, &self.inner[current..])?;
         match self.consume(len_consumed) {
             Ok(()) => Ok((value, len_consumed)),
             Err(e) => Err(ParseError::LengthError(e)),
         }
     }
+}
 
-    pub(crate) fn write<T: DeParse>(&mut self, header: &T) -> Result<(), DeParseError<T::Error>> {
-        #[allow(unsafe_code)] // TODO: document safety requirements
-        let len_written = header.write(unsafe { self.inner.as_mut() })?;
-        self.consume(len_written).map_err(DeParseError::LengthError)
+impl Writer<'_> {
+
+    pub(crate) fn new(buf: &mut [u8]) -> Writer {
+        let len = buf.len();
+        Writer {
+            inner: buf,
+            remaining: len,
+        }
     }
+
+    fn consume(&mut self, n: NonZero<usize>) -> Result<(), LengthError> {
+        if n.get() >= self.remaining {
+            return Err(LengthError {
+                expected: n,
+                actual: self.remaining,
+            });
+        };
+        self.remaining -= n.get();
+        Ok(())
+    }
+    
+    pub(crate) fn write<T: DeParse>(&mut self, val: &T) -> Result<NonZero<usize>, DeParseError<T::Error>> {
+        let current = self.inner.len() - self.remaining;
+        let consumed = val.write(&mut self.inner[current..])?;
+        self.consume(consumed).map_err(DeParseError::LengthError)?;
+        Ok(consumed)
+    }
+
 }
 
 #[derive(thiserror::Error, Debug)]
@@ -99,7 +144,7 @@ pub enum ParseError<E: core::error::Error> {
 }
 
 #[derive(thiserror::Error, Debug)]
-pub enum DeParseError<E: core::error::Error> {
+pub enum DeParseError<E> {
     #[error(transparent)]
     LengthError(LengthError),
     #[error(transparent)]
