+++ net/src/tcp/mod.rs
@@ -1,0 +1,44 @@
+use crate::parse::{DeParse, DeParseError, LengthError, Parse, ParseError};
+use etherparse::TcpHeader;
+use std::num::NonZero;
+
+#[derive(Debug, Clone, PartialEq, Eq)]
+pub struct Tcp {
+    inner: TcpHeader,
+}
+
+impl Parse for Tcp {
+    type Error = etherparse::err::tcp::HeaderSliceError;
+
+    fn parse(buf: &[u8]) -> Result<(Self, NonZero<usize>), ParseError<Self::Error>> {
+        let (inner, rest) = TcpHeader::from_slice(buf).map_err(ParseError::FailedToParse)?;
+        assert!(
+            rest.len() < buf.len(),
+            "rest.len() >= buf.len() ({rest} >= {buf})",
+            rest = rest.len(),
+            buf = buf.len()
+        );
+        let consumed = NonZero::new(buf.len() - rest.len()).ok_or_else(|| unreachable!())?;
+        Ok((Self { inner }, consumed))
+    }
+}
+
+impl DeParse for Tcp {
+    type Error = ();
+
+    fn size(&self) -> NonZero<usize> {
+        NonZero::new(self.inner.header_len()).unwrap_or_else(|| unreachable!())
+    }
+
+    fn write(&self, buf: &mut [u8]) -> Result<NonZero<usize>, DeParseError<Self::Error>> {
+        let len = buf.len();
+        if len < self.size().get() {
+            return Err(DeParseError::LengthError(LengthError {
+                expected: self.size(),
+                actual: len,
+            }));
+        };
+        buf[..self.size().get()].copy_from_slice(&self.inner.to_bytes());
+        Ok(self.size())
+    }
+}
