+++ net/src/vxlan/mod.rs
@@ -1,0 +1,3 @@
+//! VXLAN validation and manipulation tools.
+
+pub mod vni;
