+++ scripts/dpdk-sys.env
@@ -1,2 +1,2 @@
 DPDK_SYS_BRANCH="main"
-DPDK_SYS_COMMIT="e686d81e94bb8e4ca126b7ffe72ebe1afcc9e3c4"
+DPDK_SYS_COMMIT="604a812588a5f58012276d1e6b21708c424657ec"
