+++ scripts/rust.env
@@ -1,4 +1,4 @@
-LINKER="-C link-arg=-fuse-ld=lld"
+LINKER="-Clinker=./compile-env/bin/clang -C link-arg=--ld-path=./compile-env/bin/ld.lld -C link-arg=-flto"
 RELRO="-C relro-level=full"
 CRT="-C target-feature=+crt-static"
 DEBUG="-C debuginfo=full -C split-debuginfo=off"
@@ -6,16 +6,16 @@ DEBUG_ASSERTIONS_ON="-C debug-assertions=on"
 DEBUG_ASSERTIONS_OFF="-C debug-assertions=off"
 OVERFLOW_CHECK_ON="-C overflow-checks=on"
 OVERFLOW_CHECK_OFF="-C overflow-checks=off"
-LTO="-C linker-plugin-lto -C lto=thin -C embed-bitcode=yes -C codegen-units=1 -C link-arg=-Wl,--plugin-opt=-lto-embed-bitcode=optimized -C link-arg=-Wl,-u_exit -C link-arg=-Wl,-u_init -C link-arg=-Wl,-usetenv -C link-arg=-Wl,-ugetenv -C link-arg=-Wl,-u__errno_location -C link-arg=-Wl,-urecv -C link-arg=-Wl,-upoll -C link-arg=-Wl,-usleep -C link-arg=-Wl,-uusleep -C link-arg=-Wl,-unanosleep -C link-arg=-Wl,-upthread_create -C link-arg=-Wl,-upthread_key_create -C link-arg=-Wl,-upthread_setspecific -C link-arg=-Wl,-upthread_getspecific -C link-arg=-Wl,-upthread_self -C link-arg=-Wl,-upthread_getattr_np -C link-arg=-Wl,-upthread_attr_destroy -C link-arg=-Wl,-upthread_attr_setstack -C link-arg=-Wl,-upthread_attr_getstack -C link-arg=-Wl,-upthread_attr_getstacksize -C link-arg=-Wl,-upthread_attr_setstacksize -C link-arg=-Wl,-upthread_cancel -C link-arg=-Wl,-u_pthread_cleanup_push -C link-arg=-Wl,-u_pthread_cleanup_pop -C link-arg=-Wl,-u_pthread_cleanup_push -C link-arg=-Wl,-u_pthread_cleanup_pop -C link-arg=-Wl,-upthread_mutex_lock -C link-arg=-Wl,-upthread_mutex_trylock -C link-arg=-Wl,-upthread_spin_lock -C link-arg=-Wl,-upthread_spin_lock -C link-arg=-Wl,-upthread_spin_trylock"
+LTO="-C linker-plugin-lto -C lto=thin -C embed-bitcode=yes -C codegen-units=1 -C link-arg=-Wl,--plugin-opt=-lto-embed-bitcode=optimized"
+#LTO="-C lto=thin -C codegen-units=1"
+VTUNE_OPTS="-C link-arg=-Wl,-u_exit -C link-arg=-Wl,-u_init -C link-arg=-Wl,-usetenv -C link-arg=-Wl,-ugetenv -C link-arg=-Wl,-u__errno_location -C link-arg=-Wl,-urecv -C link-arg=-Wl,-upoll -C link-arg=-Wl,-usleep -C link-arg=-Wl,-uusleep -C link-arg=-Wl,-unanosleep -C link-arg=-Wl,-upthread_create -C link-arg=-Wl,-upthread_key_create -C link-arg=-Wl,-upthread_setspecific -C link-arg=-Wl,-upthread_getspecific -C link-arg=-Wl,-upthread_self -C link-arg=-Wl,-upthread_getattr_np -C link-arg=-Wl,-upthread_attr_destroy -C link-arg=-Wl,-upthread_attr_setstack -C link-arg=-Wl,-upthread_attr_getstack -C link-arg=-Wl,-upthread_attr_getstacksize -C link-arg=-Wl,-upthread_attr_setstacksize -C link-arg=-Wl,-upthread_cancel -C link-arg=-Wl,-u_pthread_cleanup_push -C link-arg=-Wl,-u_pthread_cleanup_pop -C link-arg=-Wl,-u_pthread_cleanup_push -C link-arg=-Wl,-u_pthread_cleanup_pop -C link-arg=-Wl,-upthread_mutex_lock -C link-arg=-Wl,-upthread_mutex_trylock -C link-arg=-Wl,-upthread_spin_lock -C link-arg=-Wl,-upthread_spin_lock -C link-arg=-Wl,-upthread_spin_trylock"
 OPTIMIZE_OFF="-C opt-level=0 ${DEBUG_ASSERTIONS_ON} ${OVERFLOW_CHECK_ON}"
 OPTIMIZE_ON="-C opt-level=3 ${LTO} ${DEBUG_ASSERTIONS_OFF} ${OVERFLOW_CHECK_OFF}"
 TARGET_CPU_DEBUG="-C target-cpu=generic"
-TARGET_CPU_RELEASE="-C target-cpu=native"
-
-
+TARGET_CPU_RELEASE="-C target-cpu=x86-64-v4"
 
 COMMON="${LINKER} ${RELRO} ${CRT} ${DEBUG}"
-RUSTFLAGS_DEBUG="${COMMON} ${OPTIMIZE_OFF} ${TARGET_CPU_DEBUG}"
-#RUSTFLAGS_RELEASE="${COMMON} ${OPTIMIZE_ON} ${TARGET_CPU_RELEASE} -Cprofile-generate"
-RUSTFLAGS_RELEASE="${COMMON} ${OPTIMIZE_ON} ${TARGET_CPU_RELEASE} -Cprofile-use=/tmp/pgo-data/merged.profdata"
+RUSTFLAGS_DEBUG="${COMMON} ${OPTIMIZE_OFF} ${TARGET_CPU_DEBUG} ${RUSTFLAGS_EXTRA}"
+RUSTFLAGS_RELEASE="${COMMON} ${OPTIMIZE_ON} ${TARGET_CPU_RELEASE} ${RUSTFLAGS_EXTRA}"
+#RUSTFLAGS_RELEASE="${COMMON} ${OPTIMIZE_ON} ${TARGET_CPU_RELEASE} -Cprofile-use=/tmp/pgo-data/merged.profdata"
